requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-google-genai
python-dotenv
